requirements.txt
asyncio
websocket